% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodc_classes.R
\name{bodc_classes}
\alias{bodc_classes}
\title{List unique classes or properties in the triplestore}
\usage{
bodc_classes(nerc = TRUE, instances = FALSE)
}
\arguments{
\item{nerc}{: boolean parameter. whether the endpoint queried is nerc vocabulary server (TRUE) or bodc (FALSE)}

\item{instances}{: boolean parameter. whether all unique instances of each class should be returned
we recommand you first check the number of instances in each class, this could take a long time!}
}
\description{
The function searches the triplestore for classes and properties.
The classes and properties in RDF datasets can be declared but that
is not always the case
}

