% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodc_get_triples.R
\name{bodc_get_triples}
\alias{bodc_get_triples}
\title{Access RDF triples at nerc or bodc sparql endpoints}
\usage{
bodc_get_triples(mylimit, nerc = FALSE, named_graph = FALSE)
}
\arguments{
\item{mylimit}{is the number of triples to extract. 
It can be string or integer (the function lets you use numeric and converts it internally). 
Note that this operation can take a while if mylimit is high (>1e4). We recommend you do not try 
to visualize all triples at once}

\item{nerc}{is boolean TRUE/FALSE with default TRUE. 
If TRUE (the default), access triples from nerc vocab endpoint. 
If FALSE, triples at BODC enpoint are queried.}

\item{named_graph}{is boolean TRUE/FALSE with default FALSE. 
Named graphs are sets of triples, rather than individual triples}
}
\description{
The function makes use of one of two sparql endpoints that can be used together 
to access BODC data. The first endpoint is the nerc vocab.
It contains the list of appropriate terms needed to query the bodc endpoint.  
the function returns a 3-column table with names s, p and o for subject, predicate and object
this is the typical "triple" format of RDFs, the building block of linked data
}

