% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cefas_data.R
\name{download_cefas_data}
\alias{download_cefas_data}
\title{Download CEFAS data using the CEFAS API}
\usage{
download_cefas_data(recordset_id)
}
\arguments{
\item{recordset_id}{is the ID number of the recordset to download.}
}
\value{
a fields data.frame containg the name of all fields, an Id reference for each field,
the type of information it contains, the exact column name and information about
minimum/maximum values
}
\description{
The function returns CEFAS data as a csv file directly
loaded into R
}
\details{
\code{download_cefas_data} takes one argument recordset_id. It loads
the corresponding data into R. It takes one argument recordset_id
that identifies the data to download. The list of csv
files and their corresponding recordset ids can be found be running the
function get_cefas_datasets and extracting the to_download object.
}
\examples{
mydata <- get_cefas_datasets()
mydownloadable <- mydata$to_download
get_cefas_dataset_fields(mydownloadable$Id[1])

}

