% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_getallparam.R
\name{emodnet_getallparam}
\alias{emodnet_getallparam}
\title{Function to list the physics parameters for which EMODnet has data}
\usage{
emodnet_getallparam()
}
\description{
This function uploads EMODnet metadata about both the individual parameters
and the parameter groups for which data is available.
}
\details{
\code{getallparam} produces a list of parameters for which data is available.
It can be used in tandem with other functions that list all stations 
(potentially within a given ecoregion). The function takes no parameter and
gives an exhaustive list of parameters.
}

