% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_getallplatform.R
\name{emodnet_getallplatform}
\alias{emodnet_getallplatform}
\title{Function to list the stations for which EMODnet has physics data}
\usage{
emodnet_getallplatform()
}
\description{
This function uploads EMODnet metadata about the location of stations
as well as the physics parameters that have been monitored at these stations.
}
\details{
\code{emodnet_getallplatform} produces a list of platform at which data has been monitored.
The function takes no parameter and gives an exhaustive list of stations.
}

