% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_getdata_atplatform.R
\name{emodnet_getdata_atplatform}
\alias{emodnet_getdata_atplatform}
\title{Fetch recent raw EMODnet data}
\usage{
emodnet_getdata_atplatform(platformid, paramcode)
}
\arguments{
\item{platformid}{is the EMODnet specific code for the
location at which the data is recorded. The full list
of platform ids is provided by the function \code{emodnet_getallplatform}}

\item{paramcode}{is the EMODnet specific code for the
parameter of interest. The full list
of parameters and their corresponding codes is provided 
by the function \code{emodnet_getallparam}}
}
\value{
a data.frame containing the requested data.
}
\description{
The function returns recent (up to two months old)
high-resolution time series from the EMODnet physics portal, 
given a location and a parameter.
}
\examples{
library(XML)
library(RCurl)
library(merpWS)
library(magrittr)
# get platform ids and parameter codes
list_platform <- emodnet_getallplatform()
list_param <- emodnet_getallparam()

# get current data for one platform where temperature data is available
myresult <- emodnet_getdata_atplatform(platformid = 437, paramcode = "TEMP")
}

