% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cefas_datasets.R
\name{get_cefas_datasets}
\alias{get_cefas_datasets}
\title{List CEFAS data sources availabe from CEFAS API}
\usage{
get_cefas_datasets()
}
\value{
two data.frames data_sources (integral list of recordsets)
and to_download (list of all csv files)
}
\description{
The function returns all recordsets as well as a
subset of recordsets that can be downloaded (i.e. csv files)
using the API.
}
\details{
\code{get_cefas_datasets} takes no argument. The original API returns
 a list of csv files and pdfs, not all of them can be downloaded. This functions
 returns both the full set of recordsets and a subset containing all csv files
 which in principle can all be downloaded. Note that this is a smallest
 set of datasets that can be accessed from the CEFAS data hub.
}
\examples{
mydata <- get_cefas_datasets()


}

