% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_datras_data.R
\name{get_datras_data}
\alias{get_datras_data}
\title{Function to upload DATRAS survey data into R}
\usage{
get_datras_data(whatinfo, survey, year, quarter)
}
\arguments{
\item{whatinfo}{is a character variable "HH", "HL" or "CA".
"HH" is haul information, "HL" is haul contents by size, and "CA" is haul contents by age}

\item{survey}{is a character variable matching the survey code wanted}

\item{year}{is a character variable for the year wanted}

\item{quarter}{is a character variable for the quarter wanted}
}
\description{
This function uploads data from one quarter and one year at a time, for a given survey.
Several other tools are available to the R user to perform this task, including
the DATRAS R package. A more sophisticated approach has been developed recently by Scott Large
and is available at http://www.ices.dk/marine-data/Documents/DATRAS/DATRAS_R.pdf.
Currently this code only exists as a pdf and would require a complete
re-write as an R script.
}
\details{
\code{get_datras_data} achieves the same result as Scott Large's code,
but leaves the task of aggregating surveys (if multiple surveys are uploaded) to the user.
Uploading the data is a slow process and partly depends on internet speed. Parallelising
the code as done by Scott Large will potentially speed things up, although available
bandwith will always be a problem. Note that a list of available surveys, years and quarters is stored
in the data.frame "datras_surveys", which can be looked up using data(datras_surveys)
}
\examples{
data(datras_surveys)
mysurvey <- datras_surveys$survey[1]
myear <- datras_surveys$year[1]
myquarter <- datras_surveys$quarter[1]
get_datras_data("HH", mysurvey, myear, myquarter)
}

