% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getallplatformroos.R
\name{getallplatformroos}
\alias{getallplatformroos}
\title{Function to list the stations for which EMODnet has physics data within a given region}
\usage{
getallplatformroos(roosid)
}
\arguments{
\item{roosid}{is a character variable "1" to "6". 
BOOS is Baltic, NOOS is north West European Shelves ("3"), IBI is Irland Biscay Iberia, MOON mediterranean}
}
\description{
This function uploads EMODnet metadata about the location of stations
as well as the physics parameters that have been monitored at these stations. It 
differs from getallplatform because it filters information by ecoregions or ROOS
i.e. Regional Operational Oceanographic System
}
\details{
\code{getallplatformroos} produces a list of platform at which data has been monitored 
within a given region.
}

