% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_diversity.R
\name{grid_diversity}
\alias{grid_diversity}
\title{Map diversity on lattice}
\usage{
grid_diversity(mydata, taxonomic_level, myresolution, myzoom,
  lat_centre = NULL, lon_centre = NULL)
}
\arguments{
\item{path_to_manuela}{is the path to the local}

\item{path_to_manuela}{is the path to the local}

\item{path_to_manuela}{is the path to the local}

\item{path_to_manuela}{is the path to the local}

\item{path_to_manuela}{is the path to the local}
}
\description{
This functions subsets data
}
\details{
\code{grid_diversity} merges 4 of the tables together.
The database itself is on Strathcloud in module1/data
and needs to be downloaded on the user's computer.
}
\examples{
manuela <- manueladb_to_table("~/manuela")
class(manuela)
}

