% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manueladb_to_table.R
\name{manueladb_to_table}
\alias{manueladb_to_table}
\title{Convert MANUELA to table}
\usage{
manueladb_to_table(path_to_manuela)
}
\arguments{
\item{path_to_manuela}{is the path to the local
copy of the database on the user's computer.}
}
\description{
The function takes the meiofauna database MANUELA
and uses dplyr to merge its tables into a data.frame and/or csv file.
There are 14 tables in total, some of them containing 
biotic variables and others abiotic variables.
}
\details{
\code{manueladb_to_table} currently merges 4 of the tables together.
The result is a data frame containing all biotic observations
as well as their accompanying spatial and temporal information.
The database itself is on Strathcloud in module1/data
and needs to be downloaded on the user's computer. WE HAVE BEEN 
ASKED BY THE DATA PROVIDER TO NOT SHARE IT BEYOND THE MERP COMMUNITY.
}
\examples{
manuela <- manueladb_to_table("~/manuela")
class(manuela)
}

