% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_env.R
\name{match_env}
\alias{match_env}
\title{Match biotic and abiotic data in space}
\usage{
match_env(env_lon, env_lat, bio_lon, bio_lat, env_variable = NULL)
}
\arguments{
\item{env_lon}{is}

\item{env_lat}{is}

\item{bio_lon}{is}

\item{bio_lat}{is}

\item{env_variable}{is}
}
\description{
The function takes the meiofauna database MANUELA
and use dplyr to merge its tables into a data.frame and/or csv file.
There are 14 tables in total. the first 4 concern abiotic measurements 
5 and 6 have biotic data, with names, count, length, width and biomass
do I need to disconnect the link between R and the database?
\code{match_env} merges 4 of the tables together.
The database itself is on Strathcloud in module1/data
and needs to be downloaded on the user's computer.
}
\examples{
data(datras_surveys)
}

