% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_distr.R
\name{temperature_distr}
\alias{temperature_distr}
\title{Plot species thermal niche}
\usage{
temperature_distr(fgroup = NULL, species = NULL, xtop = 1000,
  xbottom = 0)
}
\arguments{
\item{fgroup}{is a subsetting argument for the function. If non null
the function subsets the data to only retain occurrences for a particular
functional group.}

\item{species}{is a subsetting argument for the function. If non null
the function subsets the data to only retain occurrences for a particular
species.}

\item{xtop}{is the upper limit for the x axis in the thermal distribution graph}

\item{xbottom}{is the lower limit for the x axis in the thermal distribution graph}
}
\description{
The function plots the thermal distribution of species
occurrences.
}
\details{
\code{temperature_distr} takes plotting arguments and 
a data set of geolocated species occurrences with their 
accompanying temperature and taxonomy data.
}
\examples{
manuela <- manueladb_to_table("~/manuela")
class(manuela)

}

