% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datras_data.R
\name{get_datras_data}
\alias{get_datras_data}
\title{Function to upload DATRAS survey data into R}
\usage{
get_datras_data(whatinfo, survey, year, quarter)
}
\arguments{
\item{whatinfo}{is a character variable "HH", "HL" or "CA".
"HH" is haul information, "HL" is haul contents by size, and "CA" is haul contents by age}

\item{survey}{is a character variable matching the survey code wanted}

\item{year}{is a character variable for the year wanted}

\item{quarter}{is a character variable for the quarter wanted}
}
\description{
This function uploads DATRAS data from one quarter and one year at a time, for a given survey.
Several other tools are available to the R user to perform this task, including
the DATRAS R package. A more sophisticated approach has been developed recently by Scott Large
and is available in R package rICES at https://github.com/ices-tools-dev/rICES.
}
\details{
\code{get_datras_data} achieves the same result as rICES,
but leaves the task of aggregating surveys (if multiple surveys are uploaded) to the user. 
rICES makes the most of multi-core computers and will likely be faster
in downloading data for many quarters and years, but may be a little more
complex to use for an R beginner.
}
\examples{
data(datras_surveys)
mysurvey <- datras_surveys$survey[1]
myear <- datras_surveys$year[1]
myquarter <- datras_surveys$quarter[1]
get_datras_data("HH", mysurvey, myear, myquarter)
}

