% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_data.R
\name{grid_data}
\alias{grid_data}
\title{Function to grid geolocated observations on a map}
\usage{
grid_data(lat, lon, myresolution = 0.5, myzoom = 7, lat_centre = NULL,
  lon_centre = NULL)
}
\arguments{
\item{lat}{is a numerical vector of latitudes for the set of observations.}

\item{lon}{is a numerical vector of longitudes for the set of observations.}

\item{myresolution}{is the size of the cells the data is to be aggregated over}

\item{myzoom}{is the zoom to be applied to plot the gridded data on ggmap map}

\item{lat_centre}{is the user-defined latitude the map will be centred on. Default to mean 
latitude of observations.}

\item{lon_centre}{is the user-defined longitude the map will be centred on. Default to mean 
longitude of observations.}
}
\value{
\code{grid_data} returns a data frame containing the number of observations per cell.
}
\description{
Data portals such as OBIS typically return tables of geolocated observations upon requests.
Gridded data is often used to represent in a synthetic way this type of observations, 
but although R packages ggplot2 and ggmap are good at representing it, 
there is no built-in function to go from raw to aggregated data.
}
\examples{
library(robis)
obs <- occurrence("Gadus morhua", year = 2002)
justchecking <- grid_data(lat = obs$decimalLatitude, lon = obs$decimalLongitude, myresolution = 0.5, myzoom = 4)
# examine the data
head(justchecking$gridded_data)
# plot the data
justchecking$myplot
}

