% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manueladb_to_table.R
\name{manueladb_to_table}
\alias{manueladb_to_table}
\title{Convert MANUELA to table}
\usage{
manueladb_to_table(path_to_manuela)
}
\arguments{
\item{path_to_manuela}{is the path to the local
copy of the database on the user's computer.}
}
\description{
The function takes the meiofauna database MANUELA
and use dplyr to merge its tables into a data.frame and/or csv file.
There are 14 tables in total. the first 4 concern abiotic measurements 
5 and 6 have biotic data, with names, count, length, width and biomass
do I need to disconnect the link between R and the database?
\code{manueladb_to_table} merges 4 of the tables together.
The database itself is on Strathcloud in module1/data
and needs to be downloaded on the user's computer.
}
\examples{
manuela <- manueladb_to_table("~/manuela")
class(manuela)
}

