% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_env.R
\name{match_env}
\alias{match_env}
\title{Match biotic and abiotic data in space}
\usage{
match_env(env_lon, env_lat, bio_lon, bio_lat, env_variable = NULL)
}
\arguments{
\item{env_lon}{is a vector of longitudes for the abiotic variable}

\item{env_lat}{is a vector of latitude for the abiotic variable}

\item{bio_lon}{is the longitude of the biotic variable}

\item{bio_lat}{is the latitude of the biotic variable}

\item{env_variable}{is an optional vector containing the value of the abiotic variable}
}
\description{
The function takes two sets of geolocated data (typically one biotic
and one abiotic) and match their locations to combine the two.
\code{match_env} returns the value of the environmental variable
for the location in the abiotic set of coordinates that is closest to the 
location for the biotic record. If the vector of values for 
the abiotic variable is not provided, the function simply returns the
index of the closest location in the vector of coordinates.
}

