% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ersem_temp.R
\name{match_ersem_temp}
\alias{match_ersem_temp}
\title{Match ERSEM temperature to biotic observations}
\usage{
match_ersem_temp(ersem = NULL, lon, lat, year, month)
}
\arguments{
\item{lon}{is the path to the local}

\item{lat}{is the path to the local}

\item{year}{is the path to the local}

\item{month}{is the path to the local}
}
\description{
This functions uses ERSEM sea surface temperature
and matches it to a user-provided data set of
geolocated observations. The ERSEM data set is not included
in the package but can be downloaded freely from opec website
at https://portal.marineopec.eu/ It is montly data between 
jan 1991 and november 2013 on a lattice covering a domain 
between -19.83333  13.00000  and 40.11111 64.88889
latitude.
}
\details{
\code{grid_diversity} merges 4 of the tables together.
The database itself is on Strathcloud in module1/data
and needs to be downloaded on the user's computer.
}
\examples{
library(netcdf4)

}

