% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obis.R
\name{plot_obis}
\alias{plot_obis}
\title{Function to extract and map obis records}
\usage{
plot_obis(scientificname, year = NULL, region = NULL, myzoom = 7,
  lat_centre = NULL, lon_centre = NULL, gridded = F, myresolution = 0.5)
}
\arguments{
\item{region}{is a character string for a marine region as defined in marineregion.org}

\item{myzoom}{is the zoom to be applied to plot the gridded data on ggmap map}

\item{lat_centre}{is the user-defined latitude the map will be centred on. Default to mean 
latitude of observations.}

\item{lon_centre}{is the user-defined longitude the map will be centred on. Default to mean 
longitude of observations.}

\item{gridded}{is a logical argument setting whether to grid extracted obis data.}

\item{myresolution}{is the size of the cells the data is to be aggregated over}
}
\value{
\code{plot_obis} returns a data.frame containing the extracted obis data 
and a map showing occurrences.
}
\description{
This is a wrapper for the occurrence function in the robis package that extracts obis
data for a given taxa and plot its records on a map. The option of gridding the data
is available.
}
\examples{
library(robis)
records <- plot_obis("Asterias rubens", region = "United Kingdom Exclusive Economic Zone", myresolution = 0.5, myzoom = 5, gridded = T)
# examine the data
head(records$obis_data)
# plot the data
records$myplot
}

